+++ lib/features/onboarding/presentation/views/onboarding_screen.dart
@@ -2,6 +2,10 @@ import 'package:dutytable/features/auth/presentation/views/login_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+import '../widgets/onboarding_data.dart';
+import '../widgets/onboarding_indicator.dart';
+import '../widgets/onboarding_page_item.dart';
+
 class OnboardingScreen extends StatefulWidget {
   const OnboardingScreen({super.key});
 
@@ -12,24 +16,6 @@ class OnboardingScreen extends StatefulWidget {
 class _OnboardingScreenState extends State<OnboardingScreen> {
   final PageController _pageController = PageController();
   int _currentPage = 0;
-
-  final List<OnboardingData> _pages = [
-    OnboardingData(
-      title: '온보딩1',
-      body: "동아리 일정을 한눈에 관리하고\n모든 활동을 체계적으로 정리하세요",
-      image: 'assets/images/chat.png',
-    ),
-    OnboardingData(
-      title: '온보딩2',
-      body: "팀원과 실시간으로 소통하고\n중요한 공지를 놓치지 마세요",
-      image: 'assets/images/calendar.png',
-    ),
-    OnboardingData(
-      title: '온보딩3',
-      body: "회비 내역을 투명하게 관리하고\n모든 거래를 한눈에 확인하세요",
-      image: 'assets/images/money.png',
-    ),
-  ];
 
   @override
   void dispose() {
@@ -44,7 +30,7 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
   }
 
   void _goToNextPage() {
-    if (_currentPage < _pages.length - 1) {
+    if (_currentPage < onboardingPages.length - 1) {
       _pageController.nextPage(
         duration: const Duration(milliseconds: 300),
         curve: Curves.easeIn,
@@ -63,7 +49,6 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
     );
   }
 
-  //UI
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -71,7 +56,6 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
       body: SafeArea(
         child: Column(
           children: [
-            // 우측 상단 건너뛰기 버튼
             Padding(
               padding: const EdgeInsets.only(top: 16, right: 16),
               child: Align(
@@ -85,25 +69,21 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
                 ),
               ),
             ),
-
-            // PageView 영역
             Expanded(
               child: PageView.builder(
                 controller: _pageController,
                 onPageChanged: _onPageChanged,
-                itemCount: _pages.length,
+                itemCount: onboardingPages.length,
                 itemBuilder: (context, index) {
-                  return _buildPage(_pages[index]);
+                  return OnboardingPageItem(data: onboardingPages[index]);
                 },
               ),
             ),
-
-            // 인디케이터
-            _buildPageIndicator(),
-
+            OnboardingIndicator(
+              currentIndex: _currentPage,
+              length: onboardingPages.length,
+            ),
             const SizedBox(height: 24),
-
-            // 하단 버튼
             Padding(
               padding: const EdgeInsets.symmetric(horizontal: 24),
               child: SizedBox(
@@ -120,7 +100,7 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
                     elevation: 0,
                   ),
                   child: Text(
-                    _currentPage == _pages.length - 1 ? '시작하기' : '다음',
+                    _currentPage == onboardingPages.length - 1 ? '시작하기' : '다음',
                     style: const TextStyle(
                       fontSize: 16,
                       fontWeight: FontWeight.w600,
@@ -129,88 +109,10 @@ class _OnboardingScreenState extends State<OnboardingScreen> {
                 ),
               ),
             ),
-
             const SizedBox(height: 40),
           ],
         ),
       ),
     );
   }
-
-  Widget _buildPage(OnboardingData data) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 24),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          // 이미지
-          ClipRRect(
-            borderRadius: BorderRadius.circular(20),
-            child: Image.asset(
-              data.image,
-              width: 300,
-              height: 300,
-              fit: BoxFit.cover,
-            ),
-          ),
-
-          // const SizedBox(height: 14),
-
-          // 제목
-          Text(
-            data.title,
-            style: const TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
-            textAlign: TextAlign.center,
-          ),
-
-          const SizedBox(height: 16),
-
-          // 설명
-          Text(
-            data.body,
-            style: const TextStyle(
-              fontSize: 16,
-              color: Colors.grey,
-              height: 1.5,
-            ),
-            textAlign: TextAlign.center,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildPageIndicator() {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: List.generate(
-        _pages.length,
-        (index) => AnimatedContainer(
-          duration: const Duration(milliseconds: 300),
-          margin: const EdgeInsets.symmetric(horizontal: 4),
-          width: _currentPage == index ? 32 : 8,
-          height: 8,
-          decoration: BoxDecoration(
-            color: _currentPage == index
-                ? const Color(0xFF3F76EE)
-                : Colors.grey,
-            borderRadius: BorderRadius.circular(20),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// 온보딩 데이터 클래스
-class OnboardingData {
-  final String title;
-  final String body;
-  final String image;
-
-  OnboardingData({
-    required this.title,
-    required this.body,
-    required this.image,
-  });
 }
