+++ lib/features/onboarding/presentation/widgets/onboarding_indicator.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class OnboardingIndicator extends StatelessWidget {
+  final int currentIndex;
+  final int length;
+
+  const OnboardingIndicator({
+    super.key,
+    required this.currentIndex,
+    required this.length,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: List.generate(
+        length,
+        (index) => AnimatedContainer(
+          duration: const Duration(milliseconds: 300),
+          margin: const EdgeInsets.symmetric(horizontal: 4),
+          width: currentIndex == index ? 32 : 8,
+          height: 8,
+          decoration: BoxDecoration(
+            color: currentIndex == index
+                ? const Color(0xFF3F76EE)
+                : Colors.grey,
+            borderRadius: BorderRadius.circular(20),
+          ),
+        ),
+      ),
+    );
+  }
+}
