+++ lib/features/onboarding/presentation/widgets/onboarding_page_item.dart
@@ -1,0 +1,46 @@
+import 'package:flutter/material.dart';
+
+import 'onboarding_data.dart';
+
+class OnboardingPageItem extends StatelessWidget {
+  final OnboardingData data;
+
+  const OnboardingPageItem({super.key, required this.data});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 24),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          ClipRRect(
+            borderRadius: BorderRadius.circular(20),
+            child: Image.asset(
+              data.image,
+              width: 300,
+              height: 300,
+              fit: BoxFit.cover,
+            ),
+          ),
+          const SizedBox(height: 14),
+          Text(
+            data.title,
+            style: const TextStyle(fontSize: 22, fontWeight: FontWeight.bold),
+            textAlign: TextAlign.center,
+          ),
+          const SizedBox(height: 16),
+          Text(
+            data.body,
+            style: const TextStyle(
+              fontSize: 16,
+              color: Colors.grey,
+              height: 1.5,
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+}
